requirements.txt
discord.py
nrclex
nltk